refactor: all importer implementation private
