+++ cognite/neat/rules/importer/_graph2rules.py
@@ -15,7 +15,7 @@ from rdflib import Graph, Literal, Namespace, URIRef
 from cognite.neat.constants import PREFIXES
 from cognite.neat.rules import exceptions
 from cognite.neat.rules.importer._base import BaseImporter
-from cognite.neat.rules.importer.owl2rules import _create_default_metadata_parsing_config
+from cognite.neat.rules.importer._owl2rules import _create_default_metadata_parsing_config
 from cognite.neat.rules.models.tables import Tables
 from cognite.neat.utils.utils import get_namespace, remove_namespace, uri_to_short_form
 
@@ -38,11 +38,7 @@ class GraphImporter(BaseImporter):
 
     """
 
-    def __init__(
-        self,
-        graph: Graph,
-        max_number_of_instance: int = -1,
-    ):
+    def __init__(self, graph: Graph, max_number_of_instance: int = -1):
         self.graph = graph
         self.max_number_of_instance = max_number_of_instance
 
@@ -79,17 +75,7 @@ def _create_default_properties_parsing_config() -> dict[str, tuple[str, ...]]:
 
 def _create_default_classes_parsing_config() -> dict[str, tuple[str, ...]]:
     # TODO: these are to be read from Class pydantic model
-    return {
-        "header": (
-            "Class",
-            "Description",
-            "Parent Class",
-            "Source",
-            "Source Entity Name",
-            "Match Type",
-            "Comment",
-        )
-    }
+    return {"header": ("Class", "Description", "Parent Class", "Source", "Source Entity Name", "Match Type", "Comment")}
 
 
 def _parse_prefixes_df(prefixes: dict[str, Namespace]) -> pd.DataFrame:
@@ -239,8 +225,7 @@ def _graph_to_data_model_dict(graph: Graph, max_number_of_instance: int = -1) ->
 
                     warnings.warn(
                         exceptions.GraphClassPropertyMultiOccurrence(
-                            class_name=class_name,
-                            property_name=property_name,
+                            class_name=class_name, property_name=property_name
                         ).message,
                         category=exceptions.GraphClassPropertyMultiOccurrence,
                         stacklevel=3,
