+++ pyproject.toml
@@ -19,8 +19,6 @@ exclude = [
     "cognite/neat/app/ui/neat-app/node_modules/**",
     "cognite/neat/app/ui/neat-app/*.json",
 ]
-[tool.poetry.group.docs.dependencies]
-mkdocs-autorefs = "^0.5.0"
 
 
 [tool.black]
@@ -92,6 +90,7 @@ mkdocs-git-authors-plugin = { version = "*", optional = true }
 mkdocs-gitbook = { version = "*", optional = true }
 mkdocs-glightbox = { version = "*", optional = true }
 pymdown-extensions = { version = "*", optional = true }
+mkdocs-autorefs = {version = "^0.5.0", optional = true}
 mkdocstrings = { version = "*", optional = true, extras = ["python"] }
 jinja2 = { version = "^3.1.2", optional = true }
 
@@ -106,6 +105,7 @@ docs = [
     "mkdocs-glightbox",
     "pymdown-extensions",
     "mkdocstrings",
+    "mkdocs-autorefs",
 ]
 google = ["gspread", "google-api-python-client", "google-auth-oauthlib"]
 excel = ["openpyxl"]
