+++ tests/rules/exporter/test_rules2pydantic_models.py
@@ -5,7 +5,7 @@ from yaml import safe_load
 
 from cognite.neat.rules import examples, exceptions
 from cognite.neat.rules.exporter.rules2pydantic_models import rules_to_pydantic_models
-from cognite.neat.rules.importer.dms2rules import DMSImporter
+from cognite.neat.rules.importer._dms2rules import DMSImporter
 
 
 def test_rules2pydantic_models(dms_compliant_rules, source_knowledge_graph):
@@ -22,10 +22,7 @@ def test_rules2pydantic_models(dms_compliant_rules, source_knowledge_graph):
 
     assert instance.external_id == "2dd9019e-bdfb-11e5-94fa-c8f73332c8f4"
     assert instance.name == "ARENDAL 300 A T1"
-    assert instance.class_to_asset_mapping == {
-        "metadata": ["mRID"],
-        "name": ["name", "aliasName"],
-    }
+    assert instance.class_to_asset_mapping == {"metadata": ["mRID"], "name": ["name", "aliasName"]}
 
     asset = instance.to_asset()
